♻️  Merge core tracking service in tracking service
We are working to remove core references and reduce
the number of layers in the code since.

Here, we propose to merge the three functions
of the core tracking service
directly into the tracking service.
