+++ back/libs/tracking/src/interceptors/tracking.interceptor.ts
@@ -14,15 +14,12 @@ import { LoggerService } from '@fc/logger';
 import { LoggerService as LoggerLegacyService } from '@fc/logger-legacy';
 
 import { Track } from '../decorators/track.decorator';
-import { CoreTrackingService, TrackingService } from '../services';
+import { TrackingService } from '../services';
 
 @Injectable()
 export class TrackingInterceptor implements NestInterceptor {
-  // Dependency injection can require more than 4 parameters
-  // eslint-disable-next-line max-params
   constructor(
     // Injecting the two following dependencies to make them available for the tracking service
-    private readonly appTrackingService: CoreTrackingService,
     private readonly loggerLegacy: LoggerLegacyService,
     private readonly tracking: TrackingService,
     private readonly logger: LoggerService,
