+++ back/libs/tracking/src/services/index.ts
@@ -1,2 +1,1 @@
-export * from './core-tracking.service';
 export * from './tracking.service';
