+++ back/libs/tracking/src/tracking.module.ts
@@ -2,12 +2,11 @@ import { Module } from '@nestjs/common';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 
 import { TrackingInterceptor } from './interceptors';
-import { CoreTrackingService, TrackingService } from './services';
+import { TrackingService } from './services';
 
 @Module({
   providers: [
     TrackingService,
-    CoreTrackingService,
     {
       provide: APP_INTERCEPTOR,
       useClass: TrackingInterceptor,
